Removed unnecessary method declarations from utils.js; Moved the showNotification method from utils.js to MosesService.js
